requirements.txt
edgartools==3.1.2
